+++ Patterns/FactoryMethod/IPerks.cs
@@ -2,6 +2,6 @@ namespace FactoryMethod
 {
     public interface IPerks
     {
-        double EarningBonus();
+        decimal EarningBonus();
     }
 }
